move gsd creation to a separate file
